requirements.txt
# AdvanceMarkets Terminal v3.0 — Python Backend

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client (async)
httpx>=0.27.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.1.0

# AI / LLM
google-generativeai>=0.8.0

# Caching
cachetools>=5.3.0

# Environment
python-dotenv>=1.0.0
